Save delays as floats

+++ slave.py
@@ -73,10 +73,10 @@ class Slave(threading.Thread):
         
                 min,avg,max,mdev = re.findall("(\d+\.?\d*)/(\d+\.?\d*)/(\d+\.?\d*)/(\d+\.?\d*)", output)[0]
                 self.delay_res = {
-                        'delay_min': int(min),
-                        'delay_avg': int(avg),
-                        'delay_max': int(max),
-                        'delay_mdev': int(mdev)
+                        'delay_min': float(min),
+                        'delay_avg': float(avg),
+                        'delay_max': float(max),
+                        'delay_mdev': float(mdev)
                         }
                 self.delay_finish.set()
             except KeyboardInterrupt:
